+++ TablePro/Core/Services/AppServices.swift
@@ -4,6 +4,7 @@
 //
 
 import SwiftUI
+import TableProSyncTransport
 
 @MainActor
 struct AppServices {