+++ TablePro/Core/Storage/FavoriteTablesStorage.swift
@@ -1,4 +1,5 @@
 import Foundation
+import TableProSyncTransport
 import os
 
 extension Notification.Name {