+++ TablePro/Core/Storage/GroupStorage.swift
@@ -4,6 +4,7 @@
 //
 
 import Foundation
+import TableProSyncTransport
 import os
 
 /// Service for persisting connection groups