+++ TablePro/Core/Storage/SSHProfileStorage.swift
@@ -4,6 +4,7 @@
 //
 
 import Foundation
+import TableProSyncTransport
 import os
 
 final class SSHProfileStorage {